Feature : 스크립트 형식 변경 + 테스트 코드 삭제 + CI test 주석
