--- src/components/layout/map/map.test.tsx
@@ -1,12 +1,0 @@
-import { render, screen } from '@testing-library/react'
-import Map from './map'
-
-describe('Map 렌더링 테스트', () => {
-  it('should render Map', () => {
-    render(<Map />)
-
-    const map = screen.getByText('kakaoMap')
-
-    expect(map).toBeInTheDocument
-  })
-})
