+++ src/components/layout/map/map.tsx
@@ -1,7 +1,5 @@
-import { useEffect, useRef } from 'react'
-import { useDispatch } from 'react-redux'
-import { saveLatLon } from '../../../core/redux/module/centerLatLon'
-import { saveMapObject } from '../../../core/redux/module/kakaoMap'
+import { useEffect } from 'react'
+import { useMapLoaded } from '../../../hooks/useMapLoaded'
 import CenterMoveButton from '../centerMoveButton/centerMoveButton'
 
 declare global {
@@ -10,59 +8,18 @@ declare global {
   }
 }
 
-let kakaoMap
-
 const Map = () => {
-  const kakaoMapRef = useRef<HTMLDivElement>(null)
-  const dispatch = useDispatch()
+  const { mapLoaded, onLoadKakaoMap } = useMapLoaded()
 
   useEffect(() => {
-    if (!kakaoMapRef) return
+    if (!mapLoaded) return
 
-    window.kakao.maps.load(() => {
-      const option = {
-        center: new window.kakao.maps.LatLng(
-          37.56683096014424,
-          126.97865225689458,
-        ),
-        level: 5,
-      }
-
-      kakaoMap = new window.kakao.maps.Map(kakaoMapRef.current, option)
-
-      dispatch(saveMapObject(kakaoMap))
-
-      getCenterLocation(kakaoMap)
-    })
-  }, [kakaoMapRef])
-
-  const getCenterLocation = (map: any) => {
-    // 사용자의 위치를 정상적으로 받아오면 해당 위치가 중심좌표
-    if (navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition((position) => {
-        const lat = position.coords.latitude,
-          lon = position.coords.longitude
-
-        const locPosition = new window.kakao.maps.LatLng(lat, lon)
-
-        map.setCenter(locPosition)
-
-        dispatch(saveLatLon({ Latitude: lat, Longitude: lon }))
-      })
-    }
-    // 아니라면 서울 시청이 기본 중심좌표
-    else {
-      const locPosition = new window.kakao.maps.LatLng(
-        37.56683096014424,
-        126.97865225689458,
-      )
-      map.setCenter(locPosition)
-    }
-  }
+    onLoadKakaoMap()
+  }, [mapLoaded])
 
   return (
     <>
-      <div className="MapContainer" ref={kakaoMapRef}>
+      <div className="MapContainer" id="map">
         kakaoMap
         <CenterMoveButton></CenterMoveButton>
       </div>
